requirements.txt
python-dotenv
aiohttp
tqdm